requirements.txt
fastmcp>=2.0.0
paramiko>=3.5.0,<4.0.0
tomlkit>=0.12.0
pydantic>=2.0.0
